반다라트 composable 분리
