+++ feature/home/src/main/kotlin/com/nexters/bandalart/android/feature/home/HomeScreen.kt
@@ -7,11 +7,9 @@ import android.annotation.SuppressLint
 import android.content.Context
 import android.content.Intent
 import android.widget.Toast
-import com.nexters.bandalart.android.core.ui.R
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
@@ -30,7 +28,6 @@ import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Add
 import androidx.compose.material.icons.filled.Check
 import androidx.compose.material3.Card
 import androidx.compose.material3.CardDefaults
@@ -44,29 +41,20 @@ import androidx.compose.material3.rememberModalBottomSheetState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
-import androidx.compose.runtime.saveable.rememberSaveable
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
-import androidx.compose.ui.layout.Layout
-import androidx.compose.ui.layout.layoutId
-import androidx.compose.ui.platform.LocalConfiguration
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.unit.Constraints
-import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
+import com.nexters.bandalart.android.core.ui.R
 import com.nexters.bandalart.android.core.ui.component.BandalartDeleteAlertDialog
 import com.nexters.bandalart.android.core.ui.component.BandalartDropDownMenu
-import com.nexters.bandalart.android.core.ui.component.CellText
 import com.nexters.bandalart.android.core.ui.component.EmojiText
 import com.nexters.bandalart.android.core.ui.component.FixedSizeText
 import com.nexters.bandalart.android.core.ui.component.LoadingScreen
@@ -76,18 +64,14 @@ import com.nexters.bandalart.android.core.ui.extension.nonScaleSp
 import com.nexters.bandalart.android.core.ui.extension.toColor
 import com.nexters.bandalart.android.core.ui.extension.toFormatDate
 import com.nexters.bandalart.android.core.ui.theme.Gray100
-import com.nexters.bandalart.android.core.ui.theme.Gray200
 import com.nexters.bandalart.android.core.ui.theme.Gray300
-import com.nexters.bandalart.android.core.ui.theme.Gray400
 import com.nexters.bandalart.android.core.ui.theme.Gray50
-import com.nexters.bandalart.android.core.ui.theme.Gray500
 import com.nexters.bandalart.android.core.ui.theme.Gray600
 import com.nexters.bandalart.android.core.ui.theme.Gray900
 import com.nexters.bandalart.android.core.ui.theme.MainColor
-import com.nexters.bandalart.android.core.ui.theme.White
-import com.nexters.bandalart.android.feature.home.model.BandalartCellUiModel
 import com.nexters.bandalart.android.feature.home.model.BandalartDetailUiModel
 import com.nexters.bandalart.android.feature.home.model.UpdateBandalartEmojiModel
+import com.nexters.bandalart.android.feature.home.ui.BandalartChart
 import com.nexters.bandalart.android.feature.home.ui.BandalartEmojiPicker
 import com.nexters.bandalart.android.feature.home.ui.BandalartListBottomSheet
 import com.nexters.bandalart.android.feature.home.ui.BandalartSkeleton
@@ -550,319 +534,6 @@ private fun updateBandalartListTitles(
       )
     } else {
       item
-    }
-  }
-}
-
-data class SubCell(
-  val rowCnt: Int,
-  val colCnt: Int,
-  val subCellRowIndex: Int,
-  val subCellColIndex: Int,
-  val bandalartChartData: BandalartCellUiModel,
-)
-
-@Composable
-private fun BandalartChart(
-  modifier: Modifier = Modifier,
-  bandalartKey: String,
-  uiState: HomeUiState,
-  themeColor: ThemeColor,
-  bottomSheetDataChanged: (Boolean) -> Unit,
-) {
-  val context = LocalContext.current
-  val screenWidthDp = LocalConfiguration.current.screenWidthDp.dp
-  val paddedMaxWidth = remember(screenWidthDp) {
-    screenWidthDp - (15.dp * 2)
-  }
-
-  val subCellList = listOf(
-    SubCell(2, 3, 1, 1, uiState.bandalartCellData!!.children[0]),
-    SubCell(3, 2, 1, 0, uiState.bandalartCellData.children[1]),
-    SubCell(3, 2, 1, 1, uiState.bandalartCellData.children[2]),
-    SubCell(2, 3, 0, 1, uiState.bandalartCellData.children[3]),
-  )
-
-  Layout(
-    modifier = Modifier
-      .fillMaxWidth()
-      .clip(RoundedCornerShape(8.dp)),
-    content = {
-      for (index in subCellList.indices) {
-        Box(
-          modifier
-            .layoutId(context.getString(R.string.homescreen_layout_id_text, index + 1))
-            .clip(RoundedCornerShape(12.dp))
-            .background(color = Gray100),
-          content = {
-            CellGrid(
-              bandalartKey = bandalartKey,
-              themeColor = themeColor,
-              subCell = subCellList[index],
-              rows = subCellList[index].rowCnt,
-              cols = subCellList[index].colCnt,
-              bottomSheetDataChanged = bottomSheetDataChanged,
-            )
-          },
-        )
-      }
-      Box(
-        modifier
-          .layoutId(context.getString(R.string.homescreen_main_id_text))
-          .clip(RoundedCornerShape(10.dp))
-          .background(color = (uiState.bandalartCellData.mainColor?.toColor() ?: MainColor)),
-        content = {
-          Cell(
-            isMainCell = true,
-            themeColor = themeColor,
-            cellData = uiState.bandalartCellData,
-            bandalartKey = bandalartKey,
-            bottomSheetDataChanged = bottomSheetDataChanged,
-          )
-        },
-      )
-    },
-  ) { measurables, constraints ->
-    val sub1 = measurables.first { it.layoutId == context.getString(R.string.homescreen_sub1_id_text) }
-    val sub2 = measurables.first { it.layoutId == context.getString(R.string.homescreen_sub2_id_text) }
-    val sub3 = measurables.first { it.layoutId == context.getString(R.string.homescreen_sub3_id_text) }
-    val sub4 = measurables.first { it.layoutId == context.getString(R.string.homescreen_sub4_id_text) }
-    val main = measurables.first { it.layoutId == context.getString(R.string.homescreen_main_id_text) }
-
-    val chartWidth = paddedMaxWidth.roundToPx()
-    val mainWidth = chartWidth / 5
-    val padding = 1.dp.roundToPx()
-
-    val mainConstraints = Constraints.fixed(width = mainWidth, height = mainWidth)
-    val sub1Constraints = Constraints.fixed(width = mainWidth * 3 - padding, height = mainWidth * 2 - padding)
-    val sub2Constraints = Constraints.fixed(width = mainWidth * 2 - padding, height = mainWidth * 3 - padding)
-    val sub3Constraints = Constraints.fixed(width = mainWidth * 2 - padding, height = mainWidth * 3 - padding)
-    val sub4Constraints = Constraints.fixed(width = mainWidth * 3 - padding, height = mainWidth * 2 - padding)
-
-    val mainPlaceable = main.measure(mainConstraints)
-    val sub1Placeable = sub1.measure(sub1Constraints)
-    val sub2Placeable = sub2.measure(sub2Constraints)
-    val sub3Placeable = sub3.measure(sub3Constraints)
-    val sub4Placeable = sub4.measure(sub4Constraints)
-
-    layout(width = chartWidth, height = chartWidth) {
-      mainPlaceable.place(x = mainWidth * 2, y = mainWidth * 2)
-      sub1Placeable.place(x = 0, y = 0)
-      sub2Placeable.place(x = mainWidth * 3 + padding, y = 0)
-      sub3Placeable.place(x = 0, y = mainWidth * 2 + padding)
-      sub4Placeable.place(x = mainWidth * 2 + padding, y = mainWidth * 3 + padding)
-    }
-  }
-}
-
-@Composable
-fun CellGrid(
-  bandalartKey: String,
-  themeColor: ThemeColor,
-  subCell: SubCell,
-  rows: Int,
-  cols: Int,
-  bottomSheetDataChanged: (Boolean) -> Unit,
-) {
-  Column(
-    modifier = Modifier.fillMaxSize(),
-    horizontalAlignment = Alignment.CenterHorizontally,
-    verticalArrangement = Arrangement.SpaceEvenly,
-  ) {
-    var taskIndex = 0
-    repeat(rows) { rowIndex ->
-      Row(
-        modifier = Modifier.fillMaxWidth(),
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.SpaceEvenly,
-      ) {
-        repeat(cols) { colIndex ->
-          val isSubCell = rowIndex == subCell.subCellRowIndex && colIndex == subCell.subCellColIndex
-          Cell(
-            modifier = Modifier.weight(1f),
-            bandalartKey = bandalartKey,
-            themeColor = themeColor,
-            isMainCell = false,
-            cellInfo = CellInfo(
-              isSubCell = isSubCell,
-              colIndex = colIndex,
-              rowIndex = rowIndex,
-              colCnt = cols,
-              rowCnt = rows,
-            ),
-            cellData = if (isSubCell) subCell.bandalartChartData
-            else subCell.bandalartChartData.children[taskIndex++],
-            bottomSheetDataChanged = bottomSheetDataChanged,
-          )
-        }
-      }
-    }
-  }
-}
-
-data class CellInfo(
-  val isSubCell: Boolean = false,
-  val colIndex: Int = 2,
-  val rowIndex: Int = 2,
-  val colCnt: Int = 1,
-  val rowCnt: Int = 1,
-)
-
-@Composable
-fun Cell(
-  modifier: Modifier = Modifier,
-  bandalartKey: String,
-  themeColor: ThemeColor,
-  isMainCell: Boolean,
-  cellInfo: CellInfo = CellInfo(),
-  cellData: BandalartCellUiModel,
-  bottomSheetDataChanged: (Boolean) -> Unit,
-  outerPadding: Dp = 3.dp,
-  innerPadding: Dp = 2.dp,
-  mainCellPadding: Dp = 1.dp,
-) {
-  val context = LocalContext.current
-  var openBottomSheet by rememberSaveable { mutableStateOf(false) }
-  val backgroundColor = when {
-    isMainCell -> themeColor.mainColor.toColor()
-    cellInfo.isSubCell and cellData.isCompleted -> themeColor.subColor.toColor().copy(alpha = 0.6f)
-    cellInfo.isSubCell and !cellData.isCompleted -> themeColor.subColor.toColor()
-    cellData.isCompleted -> Gray200
-    else -> White
-  }
-  Box(
-    modifier = modifier
-      .padding(
-        start = if (isMainCell) mainCellPadding else if (cellInfo.colIndex == 0) outerPadding else innerPadding,
-        end = if (isMainCell) mainCellPadding else if (cellInfo.colIndex == cellInfo.colCnt - 1) outerPadding else innerPadding,
-        top = if (isMainCell) mainCellPadding else if (cellInfo.rowIndex == 0) outerPadding else innerPadding,
-        bottom = if (isMainCell) mainCellPadding else if (cellInfo.rowIndex == cellInfo.rowCnt - 1) outerPadding else innerPadding,
-      )
-      .aspectRatio(1f)
-      .clip(RoundedCornerShape(10.dp))
-      .background(backgroundColor)
-      .clickable { openBottomSheet = !openBottomSheet },
-    contentAlignment = Alignment.Center,
-  ) {
-    // 메인 목표
-    if (isMainCell) {
-      val cellTextColor = themeColor.subColor.toColor()
-      // 메인 목표가 빈 경우
-      if (cellData.title.isNullOrEmpty()) {
-        Box(contentAlignment = Alignment.Center) {
-          Column(horizontalAlignment = Alignment.CenterHorizontally) {
-            CellText(
-              cellText = context.getString(R.string.homescreen_main_cell_text),
-              cellTextColor = cellTextColor,
-              fontWeight = FontWeight.W700,
-            )
-            Icon(
-              imageVector = Icons.Default.Add,
-              contentDescription = context.getString(R.string.add_descrption_text),
-              tint = cellTextColor,
-              modifier = Modifier
-                .size(20.dp)
-                .offset(y = (-4).dp),
-            )
-          }
-        }
-      } else {
-        CellText(
-          cellText = cellData.title,
-          cellTextColor = cellTextColor,
-          fontWeight = FontWeight.W700,
-        )
-      }
-      // 서브 목표
-    } else if (cellInfo.isSubCell) {
-      val cellTextColor = themeColor.mainColor.toColor()
-      val fontWeight = FontWeight.W700
-      // 서브 목표가 빈 경우
-      if (cellData.title.isNullOrEmpty()) {
-        Column(
-          horizontalAlignment = Alignment.CenterHorizontally,
-          verticalArrangement = Arrangement.Center,
-        ) {
-          CellText(
-            cellText = context.getString(R.string.homescreen_sub_cell_text),
-            cellTextColor = cellTextColor,
-            fontWeight = fontWeight,
-          )
-          Icon(
-            imageVector = Icons.Default.Add,
-            contentDescription = context.getString(R.string.add_descrption_text),
-            tint = cellTextColor,
-            modifier = Modifier
-              .size(20.dp)
-              .offset(y = (-4).dp),
-          )
-        }
-      } else {
-        // 서브 목표를 달성할 경우
-        CellText(
-          cellText = cellData.title,
-          cellTextColor = cellTextColor,
-          fontWeight = fontWeight,
-          textAlpha = if (cellData.isCompleted) 0.6f else 1f,
-        )
-      }
-    } else {
-      // 테스크
-      val cellTextColor = if (cellData.isCompleted) Gray400 else Gray900
-      val fontWeight = FontWeight.W500
-
-      // 테스크가 비어있는 경우
-      if (cellData.title.isNullOrEmpty()) {
-        Icon(
-          imageVector = Icons.Default.Add,
-          contentDescription = context.getString(R.string.add_descrption_text),
-          tint = Gray500,
-          modifier = Modifier.size(20.dp),
-        )
-      } else {
-        // 테스크의 목표를 달성한 경우
-        if (cellData.isCompleted) {
-          Box(
-            modifier.fillMaxSize(),
-            contentAlignment = Alignment.Center,
-          ) {
-            CellText(
-              cellText = cellData.title,
-              cellTextColor = cellTextColor,
-              fontWeight = fontWeight,
-            )
-            val image = painterResource(
-              id = com.nexters.bandalart.android.core.designsystem.R.drawable.ic_cell_check,
-            )
-            Image(
-              painter = image,
-              contentDescription = context.getString(R.string.complete_descrption_text),
-              modifier = Modifier
-                .align(Alignment.BottomEnd)
-                .offset(x = (-4).dp, y = (-4).dp),
-            )
-          }
-        } else {
-          CellText(
-            cellText = cellData.title,
-            cellTextColor = cellTextColor,
-            fontWeight = fontWeight,
-          )
-        }
-      }
-    }
-    if (openBottomSheet) {
-      BandalartBottomSheet(
-        bandalartKey = bandalartKey,
-        isSubCell = cellInfo.isSubCell,
-        isMainCell = isMainCell,
-        isBlankCell = cellData.title.isNullOrEmpty(),
-        cellData = cellData,
-        onResult = { bottomSheetState, bottomSheetDataChangedState ->
-          openBottomSheet = bottomSheetState
-          bottomSheetDataChanged(bottomSheetDataChangedState)
-        },
-      )
     }
   }
 }
