+++ feature/home/src/main/kotlin/com/nexters/bandalart/android/feature/home/util/CellInfo.kt
@@ -1,0 +1,9 @@
+package com.nexters.bandalart.android.feature.home.util
+
+data class CellInfo(
+  val isSubCell: Boolean = false,
+  val colIndex: Int = 2,
+  val rowIndex: Int = 2,
+  val colCnt: Int = 1,
+  val rowCnt: Int = 1,
+)
