+++ feature/home/src/main/kotlin/com/nexters/bandalart/android/feature/home/util/StringResource.kt
@@ -1,4 +1,4 @@
-package com.nexters.bandalart.android.feature.home.ui
+package com.nexters.bandalart.android.feature.home.util
 
 import android.content.Context
 import androidx.annotation.StringRes
